requirements.txt
# Core Frameworks
langchain
langchain-community
langchain-core
langchain-google-genai
langgraph
llama-index
llama-index-vector-stores-pinecone
llama-index-readers-google
llama-index-embeddings-huggingface
sentence-transformers
torch
transformers

# Web Search
google-api-python-client

# Google Drive Auth
google-auth-oauthlib
google-auth-httplib2

# Vector DB
pinecone-client

# Backend
fastapi
uvicorn
python-dotenv

# Utilities
pydantic
requests
